requirements.txt
flask==3.0.0
requests>=2.31.0
Pillow>=10.0.0
gunicorn>=21.2.0
